Add Agency Filter to BP List activities; Improve layout in BPList and BPListActivities
