+++ frontend/src/components/manage/Blocks/BusinessPlans/ActivitiesFilters.tsx
@@ -1,3 +1,5 @@
+import React from 'react'
+
 import {
   Checkbox,
   FormControlLabel,
@@ -20,141 +22,156 @@ export default function ActivitiesFilters(props: any) {
     form,
     handleFilterChange,
     handleParamsChange,
+    withAgency = false,
   } = props
 
   return (
     <div className="flex flex-col">
       {/*  First row */}
-      <div className="flex items-center justify-between">
-        <div className="flex flex-wrap items-center gap-3">
+      <div className="flex gap-4">
+        {/* Search Input */}
+        <Field
+          name="search"
+          FieldProps={{ className: 'mb-0 w-60 BPList' }}
+          placeholder="Search by keyword..."
+          InputProps={{
+            startAdornment: (
+              <InputAdornment position="start">
+                <MuiIconButton
+                  aria-label="search table"
+                  edge="start"
+                  tabIndex={-1}
+                  onClick={() => {
+                    const search = form.current.search.value
+                    handleParamsChange({
+                      offset: 0,
+                      search,
+                    })
+                    handleFilterChange({ search })
+                  }}
+                  disableRipple
+                >
+                  <IoSearchOutline />
+                </MuiIconButton>
+              </InputAdornment>
+            ),
+          }}
+          onKeyDown={(event: any) => {
+            const search = form.current.search.value
+            if (event.key === KEY_ENTER) {
+              handleParamsChange({
+                offset: 0,
+                search,
+              })
+              handleFilterChange({ search })
+            }
+          }}
+        />
+        {/* Rest of filters */}
+        <div className="flex h-full flex-wrap items-center gap-x-4 gap-y-2 border-0 border-l border-solid border-gray-200 pl-4">
           <Field
-            name="search"
-            placeholder="Search by keyword..."
-            InputProps={{
-              startAdornment: (
-                <InputAdornment position="start">
-                  <MuiIconButton
-                    aria-label="search table"
-                    edge="start"
-                    tabIndex={-1}
-                    onClick={() => {
-                      const search = form.current.search.value
-                      handleParamsChange({
-                        offset: 0,
-                        search,
-                      })
-                      handleFilterChange({ search })
-                    }}
-                    disableRipple
-                  >
-                    <IoSearchOutline />
-                  </MuiIconButton>
-                </InputAdornment>
-              ),
+            FieldProps={{ className: 'mb-0 w-40 BPList' }}
+            Input={{ label: 'Country' }}
+            getOptionLabel={(option: any) => option?.name}
+            options={commonSlice.countries.data}
+            value={filters.country_id}
+            widget="autocomplete"
+            onChange={(_: any, value: any) => {
+              handleFilterChange({ country_id: value })
+              handleParamsChange({
+                country_id: value.map((item: any) => item.id).join(','),
+                offset: 0,
+              })
             }}
-            onKeyDown={(event: any) => {
-              const search = form.current.search.value
-              if (event.key === KEY_ENTER) {
-                handleParamsChange({
-                  offset: 0,
-                  search,
-                })
-                handleFilterChange({ search })
-              }
-            }}
+            multiple
           />
-          <div className="w-40">
+          {withAgency && (
             <Field
-              Input={{ label: 'Country' }}
+              FieldProps={{ className: 'mb-0 w-40 BPList' }}
+              Input={{ label: 'Agency' }}
               getOptionLabel={(option: any) => option?.name}
-              options={commonSlice.countries.data}
-              value={filters.country_id}
+              options={commonSlice.agencies.data}
+              value={filters.agency_id}
               widget="autocomplete"
               onChange={(_: any, value: any) => {
-                handleFilterChange({ country_id: value })
+                handleFilterChange({ agency_id: value })
                 handleParamsChange({
-                  country_id: value.map((item: any) => item.id).join(','),
+                  agency_id: value.map((item: any) => item.id).join(','),
                   offset: 0,
                 })
               }}
               multiple
             />
-          </div>
-          <div className="w-40">
-            <Field
-              Input={{ label: 'Cluster' }}
-              getOptionLabel={(option: any) => option?.name}
-              options={clusters}
-              value={filters.project_cluster_id}
-              widget="autocomplete"
-              onChange={(_: any, value: any) => {
-                handleFilterChange({ project_cluster_id: value })
-                handleParamsChange({
-                  offset: 0,
-                  project_cluster_id: value
-                    .map((item: any) => item.id)
-                    .join(','),
-                })
-              }}
-              multiple
-            />
-          </div>
-          <div className="w-40">
-            <Field
-              Input={{ label: 'Sector' }}
-              getOptionLabel={(option: any) => option?.name}
-              options={bpSlice.sectors.data}
-              value={filters.sector_id}
-              widget="autocomplete"
-              onChange={(_: any, value: any) => {
-                handleFilterChange({ sector_id: value })
-                handleParamsChange({
-                  offset: 0,
-                  sector_id: value.map((item: any) => item.id).join(','),
-                })
-              }}
-              multiple
-            />
-          </div>
-          <div className="w-40">
-            <Field
-              Input={{ label: 'Subsector' }}
-              getOptionLabel={(option: any) => option?.name}
-              options={bpSlice.subsectors.data}
-              value={filters.subsector_id}
-              widget="autocomplete"
-              onChange={(_: any, value: any) => {
-                handleFilterChange({ subsector_id: value })
-                handleParamsChange({
-                  offset: 0,
-                  subsector_id: value.map((item: any) => item.id).join(','),
-                })
-              }}
-              multiple
-            />
-          </div>
-          <div className="w-40">
-            <Field
-              Input={{ label: 'Type' }}
-              getOptionLabel={(option: any) => option?.name}
-              options={bpSlice.types.data}
-              value={filters.project_type_id}
-              widget="autocomplete"
-              isOptionEqualToValue={(option: any, value: any) =>
-                option.id === value
-              }
-              onChange={(_: any, value: any) => {
-                handleFilterChange({ project_type_id: value })
-                handleParamsChange({
-                  offset: 0,
-                  project_type_id: value.map((item: any) => item.id).join(','),
-                })
-              }}
-              multiple
-            />
-          </div>
+          )}
+          <Field
+            FieldProps={{ className: 'mb-0 w-40 BPList' }}
+            Input={{ label: 'Cluster' }}
+            getOptionLabel={(option: any) => option?.name}
+            options={clusters}
+            value={filters.project_cluster_id}
+            widget="autocomplete"
+            onChange={(_: any, value: any) => {
+              handleFilterChange({ project_cluster_id: value })
+              handleParamsChange({
+                offset: 0,
+                project_cluster_id: value.map((item: any) => item.id).join(','),
+              })
+            }}
+            multiple
+          />
+          <Field
+            FieldProps={{ className: 'mb-0 w-40 BPList' }}
+            Input={{ label: 'Sector' }}
+            getOptionLabel={(option: any) => option?.name}
+            options={bpSlice.sectors.data}
+            value={filters.sector_id}
+            widget="autocomplete"
+            onChange={(_: any, value: any) => {
+              handleFilterChange({ sector_id: value })
+              handleParamsChange({
+                offset: 0,
+                sector_id: value.map((item: any) => item.id).join(','),
+              })
+            }}
+            multiple
+          />
+          <Field
+            FieldProps={{ className: 'mb-0 w-40 BPList' }}
+            Input={{ label: 'Subsector' }}
+            getOptionLabel={(option: any) => option?.name}
+            options={bpSlice.subsectors.data}
+            value={filters.subsector_id}
+            widget="autocomplete"
+            onChange={(_: any, value: any) => {
+              handleFilterChange({ subsector_id: value })
+              handleParamsChange({
+                offset: 0,
+                subsector_id: value.map((item: any) => item.id).join(','),
+              })
+            }}
+            multiple
+          />
+          <Field
+            FieldProps={{ className: 'mb-0 w-40 BPList' }}
+            Input={{ label: 'Type' }}
+            getOptionLabel={(option: any) => option?.name}
+            options={bpSlice.types.data}
+            value={filters.project_type_id}
+            widget="autocomplete"
+            isOptionEqualToValue={(option: any, value: any) =>
+              option.id === value
+            }
+            onChange={(_: any, value: any) => {
+              handleFilterChange({ project_type_id: value })
+              handleParamsChange({
+                offset: 0,
+                project_type_id: value.map((item: any) => item.id).join(','),
+              })
+            }}
+            multiple
+          />
           <FormControlLabel
-            className="widget"
+            className="BPList !m-0"
             label="Multi-Year"
             control={
               <Checkbox
